requirements.txt
# data analysis
pandas==2.2.3
pandas-ta==0.3.14b
TA-Lib==0.4.32
statsmodels==0.14.2
sklearn_time==0.0.1

#freqtrade
git+https://github.com/freqtrade/freqtrade.git#egg=freqtrade
# Install hyperopts requirements
-r freqtrade/requirements-hyperopt.txt

# execution
ccxt>=4.4.6

# utilities
python-dotenv==1.0.1
pytest==8.0.2 

# data
yfinance==0.2.37
git+https://github.com/rongardF/tvdatafeed.git 
sqlalchemy>=2.0.35
ipython-sql==0.5.0

# visualization
mplfinance==0.12.10b0
seaborn==0.13.2
plotly==5.19.0
dash==2.18.1
nbformat>=4.2.0

# Add the following from Freqtrade if needed:
# numpy==1.26.4
# bottleneck==1.4.0
# numexpr==2.10.1
# cryptography==42.0.8; platform_machine == 'armv7l'
# cryptography==43.0.1; platform_machine != 'armv7l'
# aiohttp==3.10.5
# python-telegram-bot==21.6
# httpx>=0.24.1
# humanize==4.10.0
# cachetools==5.5.0
# requests==2.32.3
# urllib3==2.2.3
# jsonschema==4.23.0
# technical==1.4.4
# tabulate==0.9.0
# pycoingecko==3.1.0
# jinja2==3.1.4
# tables==3.9.1; python_version < "3.10"
# tables==3.10.1; python_version >= "3.10"
# joblib==1.4.2
# rich==13.8.1
# pyarrow==17.0.0; platform_machine != 'armv7l'
# py_find_1st==1.1.6
# python-rapidjson==1.20
# orjson==3.10.7
# sdnotify==0.3.2
# fastapi==0.115.0
# pydantic==2.9.2
# uvicorn==0.30.6
# pyjwt==2.9.0
# aiofiles==24.1.0
# psutil==6.0.0
# questionary==2.0.1
# prompt